Refactor mentor related requests into a separate service
