+++ src/services/StudentService.js
@@ -25,70 +25,6 @@ export default class StudentService extends SpringBootService {
     return response;
   }
 
-  async getAllMentors() {
-    /*
-    const response = {
-      data: [
-        {
-          profilePicture: "/ződ.png",
-          userId: "1",
-          firstName: "Zöld",
-          lastName: "Zoli",
-          technologies: ["React"],
-        },
-        {
-          profilePicture: "/nari.png",
-          userId: "2",
-          firstName: "Narancs",
-          lastName: "Nándi",
-          technologies: ["React", "C#", "Javascript"],
-        },
-        {
-          profilePicture: "/feka.png",
-          userId: "3",
-          firstName: "Fekete",
-          lastName: "Feri",
-          technologies: ["Java", "Spring"],
-        },
-        {
-          profilePicture: "/sari.png",
-          userId: "4",
-          firstName: "Sárga",
-          lastName: "Sára",
-          technologies: ["C#", "ASP.Net"],
-        },
-      ],
-    }; */
-    const response = await axios.get(`${this.baseURL}/filter/get-mentors`);
-    console.log("Minden mentor", response.data);
-    return response.data;
-  }
-
-  async getFilteredMentors(techs, projects) {
-    const data = { technologyTags: techs, projectTags: projects };
-    console.log("data küldésre, küldés indul", data);
-    const response = await axios.post(
-      `${this.baseURL}/filter/get-mentors-by-tags`,
-      data
-    );
-
-    /*
-    const response = {
-      data: [
-        {
-          profilePicture: "/ződ.png",
-          userId: "1",
-          firstName: "Zöld",
-          lastName: "Zoli",
-          technologies: ["React"],
-        },
-      ],
-    };
-    */
-    console.log("response getfilteredmentorstól", response.data);
-    return response.data;
-  }
-
   getUserDataById(id) {
     return axios.get(`${this.baseURL}/user/get-user-data/${id}`);
   }
